refactor(image-service): split downloader.py into bulk-metadata + local-resolver + image-writer mixins (#804)
Decompose the 551-LOC BulkImageDownloader god-class into three
responsibility-specific mixins composed onto a slimmed concrete class,
following the package's existing mixin+protocol service pattern
(ImageService / BulkDataMixin etc.).

Behavior-preserving move-only refactor: method bodies, names, signatures
and wiring are unchanged.

- bulk_metadata.py (BulkMetadataMixin): _fetch_bulk_metadata,
  _get_cached_bulk_data_record, is_bulk_data_outdated, download_bulk_metadata.
- local_resolver.py (LocalResolverMixin): fetch_card_by_name,
  fetch_printings_by_name, _resolve_card_locally, _get_local_image_index.
- image_writer.py (ImageWriterMixin): _download_single_image,
  _download_multi_face_card, _download_single_image_multi_face,
  _download_face_asset, _build_face_filename.
- downloader_protocol.py (BulkImageDownloaderProto): shared self surface
  (cache/session/local-index state) for the _Base = Proto idiom.
- downloader.py: slimmed BulkImageDownloader composing the mixins; __init__
  keeps all instance state (session, local-index trio, cache);
  download_card_image_by_name + download_all_images orchestration; module-level
  singleton facade unchanged.

Lazy imports (schemas.BULK_DATA_CACHE, printing_index._collect_face_aliases)
moved with their methods so test monkeypatches still resolve. Class identity,
method names, and the services.image_service.__init__ re-export of
BulkImageDownloader are preserved.
